requirements.txt
mutagen>=1.45.1
Pillow